# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.BulkDeleteRequest:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  controllers.Request:
    properties:
      custom_code:
        type: string
      expiry:
        type: string
      long_url:
        type: string
    required:
    - long_url
    type: object
  controllers.UpdateURLRequest:
    properties:
      long_url:
        type: string
    required:
    - long_url
    type: object
...
  models.HTTPError:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
  models.HTTPResponseErr:
    properties:
      err:
        example: not found
        type: string
    type: object
  models.HTTPValidationError:
    properties:
      error:
        additionalProperties:
          type: string
        example:
          long_url: give valid URL
        type: object
    type: object
  models.MessageSuccess:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  models.PaginatedURLResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.URL'
        type: array
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
...
  models.PaginationMeta:
    properties:
      limit:
        example: 50
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
...
  models.URL:
    properties:
      clicks:
        type: integer
      created_at:
        type: string
      expiry:
        type: string
      id:
        description: go for uuid later
        type: integer
      long_url:
        type: string
      short_url:
        type: string
      user_id:
        type: integer
    required:
    - long_url
    - short_url
    type: object
  models.URLShortenSuccess:
    properties:
      success:
        example: url inserted to db
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: High performance URL shortener API with cache-aside caching and analytics.
  title: URL Shortener
  version: "0.67"
paths:
...
  /shorten:
    post:
      consumes:
      - application/json
      description: Shorten a long URL, optionally with a custom code and expiry
      parameters:
      - description: Shorten URL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.URLShortenSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - Bearer: []
      summary: Create a shortened URL
      tags:
      - urls
  /urls:
    get:
      consumes:
      - application/json
      description: Get a paginated list of URLs shortened by the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedURLResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPResponseErr'
      security:
      - Bearer: []
      summary: Get user's shortened URLs
      tags:
      - urls
  /urls/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a shortened URL by its ID
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPResponseErr'
      security:
      - Bearer: []
      summary: Delete a shortened URL
      tags:
      - urls
    put:
      consumes:
      - application/json
      description: Update the long URL destination of an existing shortened URL
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update URL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - Bearer: []
      summary: Update a shortened URL
      tags:
      - urls
  /urls/bulk:
    delete:
      consumes:
      - application/json
      description: Delete multiple shortened URLs by their IDs
      parameters:
      - description: Bulk delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.BulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPResponseErr'
      security:
      - Bearer: []
      summary: Bulk delete shortened URLs
      tags:
      - urls
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"